Installing Alice via HauteLookAliceBundle
To install Alice we could do it directly but instead I'm going to use an awesome bundle called HautelookAliceBundle_. Let's grab the composer require command from it's README and paste that into the terminal:

composer require hautelook/alice-bundle
This bundle is a thin layer around the Alice library, which is something that let's us load fixtures with yml files, and the same DoctrineFixturesBundle that we were talking about before. This is a really nice combination because it's going to mean that we can still run our normal php app/console doctrine:fixtures:load. But after that, instead of writing raw PHP code, all of our fixtures are going to be in these really nice yml files.

And if that doesn't sound awesome yet, just hang with me. Alice is a lot more than yml files - it contains tons of goodies.