Creating your Own Faker Formatter (Function)
So I love the random data, I love how easy this is. But one thing I don't like is that our names just aren't that realistic. We're dealing with video game heroes here and none of our names are actually of real video game heroes.

To fix this let's create our own formatter called characterName.

Now if you try this out you are going to get the error that the formatteris missing:

Unknown formatter 'characterName'