True Love with Relationships
So there's one more complication that I want to introduce, and that's relationships. I have an entity called Universe as in 'Nintendo Universe' or 'Sega Universe'.

First, let's go into our yml file and create a few of these. We'll start just like before by putting the namespace and creating a few entries under that. So I'll have one for Nintendo, one for Sega and one for classic arcade

The Character entity already has a ManyToOne_ relationship to universe on a universe property

So our goal is to take these Universe objects and set them on the charcter property.

To reference another object, just use the @ symbol and then the internal key to that object. So we'll link Mario to the Nintendo universe and everyone else, for now, to the Sega Universe.

app/console hautelook:fixtures:load

When we check it out now, sure enough we see Nintendo on top followed by 9 Segas.