Using Multiple yml Files
Because our project is pretty small I've kept everything in a single file, which I recommend that you do until it gets just too big. Once it does, feel free to separate into multiple yml files.

In our case I'll create a universe.yml file and put the universe stuff in it

Of course when you do this it's not going to work because it's only loading the characters.yml file right now. So we get a missing reference error:

Reference universe_nintendo is not defined