I don't want to brag, but these are probably the nicest super-hero fixtures ever. But we've neglected a column! The avatar.
Check out the Character entity - we have a column for this called avatarFilename:
141 lines  src/AppBundle/Entity/Character.php
... lines 1 - 10
class Character {
... lines 13 - 54
  /**
   * @ORM\Column(nullable=true)
   */
  private {dollar_sign}avatarFilename;
... lines 59 - 139
}

It's going to hold just the filename part of an image, like trogdor.png or pac-man-is-the-man.jpg. This makes sense if you look in the template for the homepage. If there's an avatarFilename, we print an img tag and expect the image to be in some uploads/avatars directory, relative to web/:
lines  app/Resources/views/Homepage/homepage.html.twig
... lines 1 - 16
  {% for character in characters %}
    <tr>
... lines 19 - 24
      <td>
... lines 26 - 31
        {% if character.avatarFilename %}
          <img src='{{ asset('uploads/avatars/'~character.avatarFilename) }}' alt='{{ character.name }}'/>
        {% endif %}
      </td>
    </tr>
... lines 37 - 45
Oh boy, so this means the avatar is a bit harder. Yea, we have to set the value in the database, but we also need to make sure to put a corresponding image file into this directory. I don't want a bunch of broken images!